DIGG-482: moving cookie banner
